Fix MS TO measurement representation
* set proper flag when saving MS Timing Offset
* use gsm_subscriber's IMSI or lchan's name if bsc_subscriber is unknown
* add comments with spec reference
* store/display MS Timing Offset instead of raw Timing Offset field from
  RSL
* Compute MS Timing Offset [-63; 192] from Timing Offset field [0; 255],
  adjust structure gsm_meas_rep with proper type to store it

Related: OS#1574

+++ openbsc/include/openbsc/meas_rep.h
@@ -39,7 +39,8 @@ struct gsm_meas_rep {
 	struct gsm_meas_rep_unidir dl;
 
 	uint8_t bs_power;
-	uint8_t ms_timing_offset;
+	/* according to 3GPP TS 48.058 § MS Timing Offset [-63; 192] */
+	int16_t ms_timing_offset;
 	struct {
 		int8_t pwr;	/* MS power in dBm */
 		uint8_t ta;	/* MS timing advance */
